Add Feed-Link and rename Feed-header to FeedHome
The FeedHome is usually the homepage of the feed-generating web-page, while
the original intent of the X-RSS-Feed field was to indicate the url of the
feed itself. That however requires some additional work to avoid sending
around passwords of authenticated feeds un-encrypted.

Also add the Title for users wanting to easily filter based on that.